+++ pkg/apis/app/v1alpha1/zz_generated.openapi.go
@@ -646,13 +646,6 @@ func schema_pkg_apis_app_v1alpha1_KogitoSupportingServiceSpec(ref common.Referen
 							},
 						},
 					},
-					"httpPort": {
-						SchemaProps: spec.SchemaProps{
-							Description: "HTTPPort will set the environment env HTTP_PORT to define which port service will listen internally.",
-							Type:        []string{"integer"},
-							Format:      "int32",
-						},
-					},
 					"propertiesConfigMap": {
 						SchemaProps: spec.SchemaProps{
 							Description: "Custom ConfigMap with application.properties file to be mounted for the Kogito service. The ConfigMap must be created in the same namespace. Use this property if you need custom properties to be mounted before the application deployment. If left empty, one will be created for you. Later it can be updated to add any custom properties to apply to the service.",