+++ pkg/infrastructure/services/prometheus.go
@@ -15,6 +15,8 @@
 package services
 
 import (
+	"net/http"
+
 	monv1 "github.com/coreos/prometheus-operator/pkg/apis/monitoring/v1"
 	"github.com/kiegroup/kogito-cloud-operator/pkg/apis/app/v1alpha1"
 	"github.com/kiegroup/kogito-cloud-operator/pkg/client"
@@ -25,7 +27,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
-	"net/http"
 )
 
 const prometheusServerGroup = "monitoring.coreos.com"
@@ -37,7 +38,7 @@ func configurePrometheus(client *client.Client, kogitoService v1alpha1.KogitoSer
 		return
 	}
 
-	deploymentAvailable, err := isDeploymentAvailable(client, kogitoService)
+	deploymentAvailable, err := IsDeploymentAvailable(client, kogitoService)
 	if err != nil {
 		return
 	}
@@ -64,7 +65,7 @@ func isPrometheusAvailable(client *client.Client) bool {
 }
 
 func isPrometheusAddOnAvailable(kogitoService v1alpha1.KogitoService) (bool, error) {
-	url := infrastructure.CreateKogitoServiceURI(kogitoService)
+	url := infrastructure.GetKogitoServiceEndpoint(kogitoService)
 	url = url + getMonitoringPath(kogitoService.GetSpec().GetMonitoring())
 	if resp, err := http.Head(url); err != nil {
 		return false, err